requirements.txt
fastapi
uvicorn[standard]
groq
pandas
python-dotenv
pydantic
scikit-learn
matplotlib
numpy
